+++ src/tests/tribol_mortar_force.cpp
@@ -74,19 +74,26 @@ public:
                            int * conn2,
                            tribol::ContactMethod method )
    {
-      if (this->numNodesPerFace != 4)
-      {
-         SLIC_ERROR("checkMortarForces: number of nodes per face not equal to 4.");
-      }
-
       // grab coordinate data
       real * x = this->x;
       real * y = this->y;
       real * z = this->z;
 
       // register the mesh with tribol
-      const int cellType = (dim == 3) ? (int)(tribol::FACE) : 
-                                        (int)(tribol::EDGE);
+      int cellType;
+      switch (this->numNodesPerFace)
+      {
+         case 4:
+         {
+            cellType = (int)(tribol::LINEAR_QUAD);      
+            break;
+         }
+         default:
+         {
+            SLIC_ERROR("checkMortarForces: number of nodes per face not equal to 4.");
+         } 
+      }
+
       const int mortarMeshId = 0;
       const int nonmortarMeshId = 1;