+++ src/tribol/common/Parameters.hpp
@@ -36,11 +36,16 @@ constexpr integer ANY_MESH = -1;
 enum InterfaceElementType
 {
    UNDEFINED_ELEMENT, ///! Undefined
-   NODE,              ///! 0D node
-   EDGE,              ///! 1D edge
-   FACE,              ///! 2D face
-   CELL,              ///! 3D volume
-   NUM_CONTACT_ELEMENTS = CELL
+//   NODE,              ///! 0D node
+//   EDGE,              ///! 1D edge
+//   FACE,              ///! 2D face
+//   CELL,              ///! 3D volume
+   LINEAR_EDGE,       ///! 1D linear edge 
+   LINEAR_TRIANGLE,   ///! 2D linear triangle
+   LINEAR_QUAD,       ///! 2D linear quadrilateral
+   LINEAR_TET,        ///! 3D linear tetrahedron (volume methods and test mesh class)
+   LINEAR_HEX,        ///! 3D linear hexahedron (volume methods and test mesh class)
+   NUM_CONTACT_ELEMENTS = LINEAR_HEX
 };
 
 /*!