+++ src/tribol/physics/Mortar.cpp
@@ -200,7 +200,6 @@ void ComputeSingleMortarGaps( CouplingScheme const * cs )
    ContactPlaneManager& cpManager = ContactPlaneManager::getInstance();
    parameters_t& parameters = parameters_t::getInstance();
    integer const dim = parameters.dimension;
-   IndexType const numNodesPerFace = (dim == 3) ? 4 : 2;
 
    ////////////////////////////////////////////////////////////////////////
    //
@@ -212,6 +211,7 @@ void ComputeSingleMortarGaps( CouplingScheme const * cs )
 
    MeshData& mortarMesh = meshManager.GetMeshInstance( mortarId );
    MeshData& nonmortarMesh = meshManager.GetMeshInstance( nonmortarId );
+   IndexType const numNodesPerFace = mortarMesh.m_numNodesPerCell;
 
    real const * const x1 = mortarMesh.m_positionX;
    real const * const y1 = mortarMesh.m_positionY; 
@@ -336,7 +336,6 @@ int ApplyNormal< SINGLE_MORTAR, LAGRANGE_MULTIPLIER >( CouplingScheme const * cs
    ContactPlaneManager& cpManager = ContactPlaneManager::getInstance();
    parameters_t& parameters = parameters_t::getInstance();
    integer const dim = parameters.dimension;
-   IndexType const numNodesPerFace = (dim == 3) ? 4 : 2;
 
    ////////////////////////////////
    //                            //
@@ -348,6 +347,7 @@ int ApplyNormal< SINGLE_MORTAR, LAGRANGE_MULTIPLIER >( CouplingScheme const * cs
 
    MeshData& mortarMesh = meshManager.GetMeshInstance( mortarId );
    MeshData& nonmortarMesh = meshManager.GetMeshInstance( nonmortarId );
+   IndexType const numNodesPerFace = mortarMesh.m_numNodesPerCell;
 
    real const * const x1 = mortarMesh.m_positionX;
    real const * const y1 = mortarMesh.m_positionY; 
@@ -738,10 +738,12 @@ int GetMethodData< MORTAR_WEIGHTS >( CouplingScheme const * cs )
    ContactPlaneManager& cpManager = ContactPlaneManager::getInstance();
    parameters_t& parameters = parameters_t::getInstance();
    integer const dim = parameters.dimension;
-   IndexType const numNodesPerFace = (dim == 3) ? 4 : 2;
 
    IndexType const mortarId = cs->getMeshId1();
    IndexType const nonmortarId = cs->getMeshId2();
+   MeshManager& meshManager = MeshManager::getInstance();
+   MeshData& mortarMesh = meshManager.GetMeshInstance( mortarId );
+   IndexType const numNodesPerFace = mortarMesh.m_numNodesPerCell;
 
    ////////////////////////////////
    //                            //